pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leasing"
version = "0.1.0"
description = "Analyse automatique de conformité Leasing Social pour HESS Automobile"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "HESS Automobile" },
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic[email]>=2.9",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
    "structlog>=24.4",
    "sentry-sdk[fastapi]>=2.18",
    "supabase>=2.10",
    "python-dotenv>=1.0",
    # Salesforce (Phase 1 — extraction des dossiers gagnés)
    "simple-salesforce>=1.12",
    # IA (Phase 1 analyse exploratoire + Phase 3 prod)
    "google-generativeai>=0.8",
    "openai>=1.55",
    # PDF / images (Phase 1 + Phase 3)
    "pypdf>=5.1",
    "pdfplumber>=0.11",
    "Pillow>=11.0",
    # Chiffrement local des PDFs téléchargés (Phase 1 — RGPD)
    "cryptography>=43.0",
    # Templates HTML (Phase 3 mail)
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7",
    "mypy>=1.13",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "types-requests",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*", "scripts*", "dataset*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
    "integration: tests d'intégration (lents, nécessitent secrets)",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*", "scripts.*"]
disallow_untyped_defs = false
